split signin, signup, forget into different js files
